Adding support for a provider
This commit adds support for a provider for this backend. I love this
backend but I really hate YAML. I also hate JSON. HOCON is a good
intermediary. I've added provider support for JSON and HOCON. This still
defaults to YAML for the provider but allows users to define a data
provider in their module's `hiera.yaml`.